Added optional output image histogram adjustment.
