+++ dev/env.yaml
@@ -14,6 +14,7 @@ dependencies:
   - opencv
   - pandas
   - pip
+  - pillow
   - python>=3.10
   - pytorch
   - requests
