+++ pyproject.toml
@@ -24,6 +24,7 @@ dependencies = [
     "numpy", 
     "opencv-python", 
     "pandas", 
+    "pillow", 
     "requests",
     "torch", 
     "torchvision",
